FastAPI_Basic_Auth/requirements.txt
fastapi==0.115.4
Jinja2==3.1.4
passlib==1.7.4
pydantic==2.9.2
pylint==3.3.2
python-dotenv==1.0.1
requests==2.32.3
uvicorn==0.32.1
